# api/openapi-full.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api-keys": {
      "post": {
        "operationId": "ApiKeysController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateApiKeyDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } }
      },
      "get": {
        "operationId": "ApiKeysController_findAll",
        "parameters": [],
        "responses": { "200": { "description": "" } }
      }
    },
    "/api-keys/{id}": {
      "get": {
        "operationId": "ApiKeysController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } }
      },
      "patch": {
        "operationId": "ApiKeysController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateApiKeyDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } }
      },
      "delete": {
        "operationId": "ApiKeysController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } }
      }
    },
    "/api-keys/{id}/roll": {
      "post": {
        "operationId": "ApiKeysController_roll",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } }
      }
    },
...
    "/customers": {
      "post": {
        "operationId": "create",
        "summary": "Create customer",
        "description": "Create customer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCustomerDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Customer has been successfully created." },
          "400": { "description": "Bad Request." }
        },
        "tags": ["Customers"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "findAllByOrganization",
        "summary": "List customers",
        "parameters": [],
        "responses": {
          "200": { "description": "List of customers." },
          "400": { "description": "Bad Request." }
        },
        "tags": ["Customers"],
        "security": [{ "bearer": [] }]
      }
    },
    "/customers/{id}": {
      "get": {
        "operationId": "findAllByOrganization",
        "summary": "List customers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Customer details." },
          "400": { "description": "Bad Request." }
        },
        "tags": ["Customers"],
        "security": [{ "bearer": [] }]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCustomerDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Customer has been successfully updated." },
          "400": { "description": "Bad Request." }
        },
        "tags": ["Customers"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "CustomersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Customer has been successfully deleted." },
          "400": { "description": "Bad Request." }
        },
        "tags": ["Customers"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/getstream/{feedName}/{feedId}": {
      "get": {
        "operationId": "GetStreamController_getFeed",
        "parameters": [
          {
            "name": "feedName",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "feedId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } }
      }
    }
  },
  "info": {
    "title": "kohortPay",
...
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    { "url": "https://api.kohortpay.dev", "description": "staging" },
    { "url": "http://localhost:3000/", "description": "Development" },
    { "url": "https://api.kohortpay.com", "description": "Production" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" },
      "basic": { "type": "http", "scheme": "basic" }
    },
    "schemas": {
...
      "CreateAddressDto": {
        "type": "object",
        "properties": {
          "addressLine1": {
            "type": "string",
            "description": "The first line of the address.",
            "example": " Avenue des Champs-Élysées"
          },
          "addressLine2": {
            "type": "string",
            "description": "The second line of the address (optional).",
            "example": "Appartement 4B",
            "nullable": true
          },
          "city": {
            "type": "string",
            "description": "The city of the address.",
            "example": "Paris"
          },
          "postalCode": {
            "type": "string",
            "description": "The postal code of the address.",
            "example": "75014"
          },
          "country": {
            "type": "string",
            "description": "The country of the address.",
            "example": "France"
          },
          "state": {
            "type": "string",
            "description": "The state of the address (optional).",
            "example": "Île-de-France",
            "nullable": true
          }
        },
        "required": ["addressLine1", "city", "postalCode", "country"]
      },
...
      "CreateApiKeyDto": { "type": "object", "properties": {} },
      "UpdateApiKeyDto": { "type": "object", "properties": {} },
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "emailAddress": {
            "type": "string",
            "description": "The primary email address of the user.",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the customer.",
            "example": "John",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the customer.",
            "nullable": true,
            "example": "Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "The phone number of the customer.",
            "nullable": true,
            "example": "+1 555 555 5555"
          },
          "address": {
            "description": "The address of the customer.",
            "allOf": [{ "$ref": "#/components/schemas/CreateAddressDto" }]
          },
          "shippingAddress": {
            "description": "The shipping address of the customer.",
            "allOf": [{ "$ref": "#/components/schemas/CreateAddressDto" }]
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata for the  customer.",
            "example": { "order_id": "ord_1JYLo8KerLxWZaQtys6ZQ1xS" }
          }
        },
        "required": ["emailAddress"]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "example": "2021-07-22T00:00:00.000Z"
          }
        }
      },
...
    }
  }
}